requirements.txt
# AWS SDK and core dependencies
boto3>=1.34.0
botocore>=1.34.0

# AWS Lambda Powertools for structured logging and tracing
aws-lambda-powertools>=2.31.0

# Type hints support
typing-extensions>=4.9.0

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.12.0

# AWS CDK dependencies for infrastructure
aws-cdk-lib>=2.100.0
constructs>=10.0.0
